backend/requirements.txt
django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
drf-spectacular==0.28.0
django-filter==24.3
django-cors-headers==4.6.0
psycopg[binary]==3.2.4
gunicorn==23.0.0
pytest==8.3.4
pytest-django==4.9.0
